moved page header component into standAloneComponents directory
